// src/shared/api/generated/__swagger_client/models/OrgUnitDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrgUnitDto = {
    id?: string;
    organizationId?: string;
    parentId?: string;
    code?: string;
    name?: string;
    unitType?: string;
    active?: boolean;
};